boostrap

+++ db/seeds.rb
@@ -7,16 +7,16 @@
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
 require_relative "../lib/api/goodapi"
-require_relative "../lib/api/aws_ecs/aws"
+#require_relative "../lib/api/aws_ecs/aws"
 
-User.delete_all
-Book.delete_all
+if User.all.length != 0 then User.delete_all end
+if Book.all.length != 0 then Book.delete_all end
 
-User.create(username: "liuchao", display_name: "Chao Liu", password: "1", admin: true, email: "lc89757@brandeis.edu")
-User.create(username: "hanzhenyu", display_name: "Zhenyu Han", password: "1", admin: true, email: "hanzhenyu@brandeis.edu")
-User.create(username: "breyerjs", display_name: "Jackson", password: "1", admin: true, email: "breyerjs@brandeis.edu")
-User.create(username: "boyang", display_name: "Boyang", password: "1", admin: true, email: "ban@brandeis.edu")
-User.create(username: "aaaaaa", password: "1", admin: false, email: "sample@sample.com")
+User.create(username: "liuchao", display_name: "Chao Liu", password: "1", admin: "1", email: "lc89757@brandeis.edu")
+User.create(username: "hanzhenyu", display_name: "Zhenyu Han", password: "1", admin: "1", email: "hanzhenyu@brandeis.edu")
+User.create(username: "breyerjs", display_name: "Jackson", password: "1", admin: "1", email: "breyerjs@brandeis.edu")
+User.create(username: "boyang", display_name: "Boyang", password: "1", admin: "1", email: "ban@brandeis.edu")
+User.create(username: "aaaaaa", password: "1", admin: "0", email: "sample@sample.com")
 Book.create(title: "Agile Web Development with Rails", author: "Sam Ruby")
 Book.create(title: "Don't Make Me Think", author: "Steve Krug")
 
@@ -27,8 +27,9 @@ Book.create(title: "Don't Make Me Think", author: "Steve Krug")
 #   Book.create(title: author, author:author)
 # end
 
-aws = Awsapi.new
-aws.search("ruby", 10)
-aws.books.each do |b|
-  book = Book.new(title:b.title, author:b.author)
-end
+# aws = Awsapi.new
+# aws.search("ruby", 10)
+# aws.books.each do |b|
+#   book = Book.new(title:b.title, author:b.author, image_url:b.image_url,description:b.description)
+#   book.save
+# end
